Aggregated factor loadings

The factor loadings aggregated over k-folds

Usage

agg_loadings(models, flag = 0.3, digits = 2)

Arguments

models

An object returned from kfa

flag

threshold below which loading will be flagged

digits

integer; number of decimal places to display in the report.

Value

data.frame of mean factor loadings for each factor model and vector with count of folds with a flagged loading

Examples

data(example.kfa)
agg_loadings(example.kfa)
#> $loadings
#> $loadings$`1-factor`
#>     variable      mean       range loading flag heywood flag
#> x1        x1 0.6774966 0.68 - 0.68            0            0
#> x2        x2 0.7798081 0.77 - 0.79            0            0
#> x3        x3 0.3057225 0.28 - 0.33            1            0
#> x4        x4 0.6829395 0.68 - 0.69            0            0
#> x5        x5 0.5862225 0.59 - 0.59            0            0
#> x6        x6 0.7768287 0.73 - 0.82            0            0
#> x7        x7 0.3346682 0.30 - 0.37            0            0
#> x8        x8 0.2889363 0.25 - 0.32            1            0
#> x9        x9 0.2770869 0.25 - 0.30            2            0
#> x10      x10 0.2089878 0.19 - 0.23            2            0
#> x11      x11 0.1586579 0.12 - 0.20            2            0
#> x12      x12 0.1963989 0.13 - 0.26            2            0
#> x13      x13 0.2663434 0.23 - 0.30            1            0
#> x14      x14 0.2827955 0.25 - 0.31            1            0
#> x15      x15 0.1658010 0.10 - 0.23            2            0
#> x16      x16 0.1467884 0.10 - 0.19            2            0
#> x17      x17 0.2390635 0.16 - 0.32            1            0
#> x18      x18 0.1217103 0.10 - 0.14            2            0
#> x19      x19 0.2078937 0.16 - 0.25            2            0
#> x20      x20 0.1188662 0.06 - 0.18            2            0
#> 
#> $loadings$`2-factor`
#>     variable      mean       range loading flag heywood flag
#> x1        x1 0.6959706 0.70 - 0.70            0            0
#> x2        x2 0.7873509 0.79 - 0.79            0            0
#> x3        x3 0.3322585 0.33 - 0.33            0            0
#> x4        x4 0.6995177 0.70 - 0.70            0            0
#> x5        x5 0.6034911 0.60 - 0.60            0            0
#> x6        x6 0.7551903 0.76 - 0.76            0            0
#> x7        x7 0.3811478 0.38 - 0.38            0            0
#> x8        x8 0.2812988 0.28 - 0.28            1            0
#> x9        x9 0.2572072 0.26 - 0.26            1            0
#> x10      x10 0.1827568 0.18 - 0.18            1            0
#> x11      x11 0.1763031 0.18 - 0.18            1            0
#> x12      x12 0.2304017 0.23 - 0.23            1            0
#> x13      x13 0.2621384 0.26 - 0.26            1            0
#> x14      x14 0.2758586 0.28 - 0.28            1            0
#> x15      x15 0.6360650 0.64 - 0.64            0            0
#> x16      x16 0.4794472 0.48 - 0.48            0            0
#> x17      x17 0.9242545 0.92 - 0.92            0            0
#> x18      x18 0.4667048 0.47 - 0.47            0            0
#> x19      x19 0.6830920 0.68 - 0.68            0            0
#> x20      x20 0.4661086 0.47 - 0.47            0            0
#> 
#> $loadings$`3-factor`
#>     variable      mean       range loading flag heywood flag
#> x1        x1 0.6934446 0.68 - 0.71            0            0
#> x2        x2 0.8083627 0.80 - 0.81            0            0
#> x3        x3 0.3099117 0.29 - 0.33            1            0
#> x4        x4 0.7002884 0.69 - 0.71            0            0
#> x5        x5 0.6070322 0.60 - 0.61            0            0
#> x6        x6 0.8090223 0.78 - 0.84            0            0
#> x7        x7 0.3432549 0.31 - 0.38            0            0
#> x8        x8 0.7898625 0.76 - 0.82            0            0
#> x9        x9 0.7403907 0.73 - 0.75            0            0
#> x10      x10 0.6148914 0.59 - 0.64            0            0
#> x11      x11 0.4599873 0.45 - 0.47            0            0
#> x12      x12 0.5117059 0.47 - 0.55            0            0
#> x13      x13 0.7119368 0.68 - 0.74            0            0
#> x14      x14 0.6338454 0.61 - 0.66            0            0
#> x15      x15 0.6216908 0.61 - 0.63            0            0
#> x16      x16 0.4820720 0.48 - 0.49            0            0
#> x17      x17 0.9089849 0.89 - 0.93            0            0
#> x18      x18 0.4491768 0.43 - 0.46            0            0
#> x19      x19 0.7095948 0.68 - 0.74            0            0
#> x20      x20 0.4883682 0.47 - 0.51            0            0
#> 
#> 
#> $flag
#> 1-factor 2-factor 3-factor 
#>        2        1        1 
#> 
#> $heywood
#> 1-factor 2-factor 3-factor 
#>        0        0        0 
#>