Aggregated scale reliabilities

The factor reliabilities aggregated over k-folds

Usage

agg_rels(models, flag = 0.6, digits = 2)

Arguments

models

An object returned from kfa

flag

threshold below which reliability will be flagged

digits

integer; number of decimal places to display in the report.

Value

data.frame of mean factor (scale) reliabilities for each factor model and vector with count of folds with a flagged reliability

Examples

data(example.kfa)
agg_rels(example.kfa)
#> Warning: 
#> The reliability() function was deprecated in 2022 and will cease to be included in future versions of semTools. See help('semTools-deprecated) for details.
#> 
#> It is replaced by the compRelSEM() function, which can estimate alpha and model-based reliability in an even wider variety of models and data types, with greater control in specifying the desired type of reliability coefficient (i.e., more explicitly choosing assumptions). 
#> 
#> The average variance extracted should never have been included because it is not a reliability coefficient. It is now available from the AVE() function.
#> Warning: 
#> The reliability() function was deprecated in 2022 and will cease to be included in future versions of semTools. See help('semTools-deprecated) for details.
#> 
#> It is replaced by the compRelSEM() function, which can estimate alpha and model-based reliability in an even wider variety of models and data types, with greater control in specifying the desired type of reliability coefficient (i.e., more explicitly choosing assumptions). 
#> 
#> The average variance extracted should never have been included because it is not a reliability coefficient. It is now available from the AVE() function.
#> Warning: 
#> The reliability() function was deprecated in 2022 and will cease to be included in future versions of semTools. See help('semTools-deprecated) for details.
#> 
#> It is replaced by the compRelSEM() function, which can estimate alpha and model-based reliability in an even wider variety of models and data types, with greater control in specifying the desired type of reliability coefficient (i.e., more explicitly choosing assumptions). 
#> 
#> The average variance extracted should never have been included because it is not a reliability coefficient. It is now available from the AVE() function.
#> Warning: 
#> The reliability() function was deprecated in 2022 and will cease to be included in future versions of semTools. See help('semTools-deprecated) for details.
#> 
#> It is replaced by the compRelSEM() function, which can estimate alpha and model-based reliability in an even wider variety of models and data types, with greater control in specifying the desired type of reliability coefficient (i.e., more explicitly choosing assumptions). 
#> 
#> The average variance extracted should never have been included because it is not a reliability coefficient. It is now available from the AVE() function.
#> Warning: 
#> The reliability() function was deprecated in 2022 and will cease to be included in future versions of semTools. See help('semTools-deprecated) for details.
#> 
#> It is replaced by the compRelSEM() function, which can estimate alpha and model-based reliability in an even wider variety of models and data types, with greater control in specifying the desired type of reliability coefficient (i.e., more explicitly choosing assumptions). 
#> 
#> The average variance extracted should never have been included because it is not a reliability coefficient. It is now available from the AVE() function.
#> $reliabilities
#> $reliabilities$`1-factor`
#>    factor omega3.mean omega3.range omega3.flag alpha.mean alpha.range
#> f1     f1   0.7907651  0.77 - 0.81           0  0.5775451 0.53 - 0.62
#>    alpha.flag
#> f1          1
#> 
#> $reliabilities$`2-factor`
#>    factor omega3.mean omega3.range omega3.flag alpha.mean alpha.range
#> f1     f1   0.8091162  0.81 - 0.81           0  0.6158628 0.62 - 0.62
#> f2     f2   0.7776355  0.78 - 0.78           0  0.7854731 0.79 - 0.79
#>    alpha.flag
#> f1          0
#> f2          0
#> 
#> $reliabilities$`3-factor`
#>    factor omega3.mean omega3.range omega3.flag alpha.mean alpha.range
#> f1     f1   0.8001072  0.79 - 0.81           0  0.8110345 0.80 - 0.82
#> f2     f2   0.8252782  0.81 - 0.84           0  0.8306236 0.81 - 0.85
#> f3     f3   0.7768455  0.78 - 0.78           0  0.7899689 0.78 - 0.80
#>    alpha.flag
#> f1          0
#> f2          0
#> f3          0
#> 
#> 
#> $flag
#> 1-factor 2-factor 3-factor 
#>        1        0        0 
#>